reorganize Gateway functions
