+++ modules/gateway/peers.go
@@ -1,6 +1,9 @@
 package gateway
 
 import (
+	"errors"
+	"math/rand"
+
 	"github.com/NebulousLabs/Sia/modules"
 )
 
@@ -9,7 +12,63 @@ const (
 	maxAddrLength = 100
 )
 
-// SharePeers returns up to 10 randomly selected peers.
+func (g *Gateway) addPeer(peer modules.NetAddress) error {
+	if _, exists := g.peers[peer]; exists {
+		return errors.New("peer already added")
+	}
+	g.peers[peer] = 0
+	return nil
+}
+
+func (g *Gateway) removePeer(peer modules.NetAddress) error {
+	if _, exists := g.peers[peer]; !exists {
+		return errors.New("no record of that peer")
+	}
+	delete(g.peers, peer)
+	return nil
+}
+
+func (g *Gateway) randomPeer() (modules.NetAddress, error) {
+	r := rand.Intn(len(g.peers))
+	for peer := range g.peers {
+		if r == 0 {
+			return peer, nil
+		}
+		r--
+	}
+	return "", ErrNoPeers
+}
+
+// AddPeer adds a peer to the Gateway's peer list.
+func (g *Gateway) AddPeer(peer modules.NetAddress) error {
+	g.mu.Lock()
+	defer g.mu.Unlock()
+	return g.addPeer(peer)
+}
+
+// RemovePeer removes a peer from the Gateway's peer list.
+func (g *Gateway) RemovePeer(peer modules.NetAddress) error {
+	g.mu.Lock()
+	defer g.mu.Unlock()
+	return g.removePeer(peer)
+}
+
+// AddPeer is an RPC that requests that the Gateway add a peer to its peer
+// list. The supplied peer is assumed to be the peer making the RPC.
+func (g *Gateway) AddMe(conn modules.NetConn) error {
+	var peer modules.NetAddress
+	err := conn.ReadObject(&peer, maxAddrLength)
+	if err != nil {
+		return err
+	}
+	if !g.Ping(peer) {
+		return ErrUnreachable
+	}
+	g.AddPeer(peer)
+	return nil
+}
+
+// SharePeers is an RPC that returns up to 10 randomly selected peers.
 func (g *Gateway) SharePeers(conn modules.NetConn) error {
 	g.mu.RLock()
 	defer g.mu.RUnlock()
@@ -40,7 +99,7 @@ func (g *Gateway) requestPeers(addr modules.NetAddress) error {
 	}
 	for _, peer := range newPeers {
 		// don't add ourselves
-		if peer == g.myAddr {
+		if peer == g.Address() {
 			continue
 		}
 		// ping each peer in a separate goroutine
@@ -51,33 +110,4 @@ func (g *Gateway) requestPeers(addr modules.NetAddress) error {
 		}(peer)
 	}
 	return nil
-}
-
-// AddPeer is an RPC that requests that the Gateway add a peer to its peer
-// list. The supplied peer is assumed to be the peer making the RPC.
-func (g *Gateway) AddMe(conn modules.NetConn) error {
-	var peer modules.NetAddress
-	err := conn.ReadObject(&peer, maxAddrLength)
-	if err != nil {
-		return err
-	}
-	if !g.Ping(peer) {
-		return ErrUnreachable
-	}
-	g.AddPeer(peer)
-	return nil
-}
-
-// AddPeer adds a peer to the Gateway's peer list.
-func (g *Gateway) AddPeer(peer modules.NetAddress) error {
-	g.mu.Lock()
-	defer g.mu.Unlock()
-	return g.addPeer(peer)
-}
-
-// RemovePeer removes a peer from the Gateway's peer list.
-func (g *Gateway) RemovePeer(peer modules.NetAddress) error {
-	g.mu.Lock()
-	defer g.mu.Unlock()
-	return g.removePeer(peer)
 }
